# src/main/java/com/example/hhplus_ecommerce/domain/dto/request/AddCartRequest.java
package com.example.hhplus_ecommerce.domain.dto.request;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Schema(description = "장바구니 추가 요청")
public class AddCartRequest {
    @Schema(description = "상품 ID", example = "1", requiredMode = Schema.RequiredMode.REQUIRED)
    private Long productId;

    @Schema(description = "수량", example = "2", requiredMode = Schema.RequiredMode.REQUIRED)
    private Integer quantity;
}